Improved update image widget - not connecting - restarting server

+++ articles/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -90,8 +91,13 @@ def article_detail(request, article_id):
     return render(request, 'articles/article_detail.html', context)
 
 
+@login_required
 def add_article(request):
     """Add new article to store"""
+
+    if not request.user.is_superuser:
+        messages.error(request, 'This area is for contributing academics only!')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ArticleForm(request.POST, request.FILES)
@@ -111,9 +117,13 @@ def add_article(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_article(request, article_id):
     """Edit article"""
+    if not request.user.is_superuser:
+        messages.error(request, 'This area is for contributing academics only!')
+        return redirect(reverse('home'))
+
     article = get_object_or_404(Article, pk=article_id)
 
     if request.method == 'POST':
@@ -136,9 +146,13 @@ def edit_article(request, article_id):
     
     return render(request, template, context)
 
-
+@login_required
 def delete_article(request, article_id):
     """Edit article"""
+    if not request.user.is_superuser:
+        messages.error(request, 'This area is for contributing academics only!')
+        return redirect(reverse('home'))
+
     article = get_object_or_404(Article, pk=article_id)
     article.delete()
     messages.success(request, 'Article deleted!')
